Refactor ApplicationController and modularize error handling
