+++ app/controllers/application_controller.rb
@@ -4,20 +4,14 @@ require 'feature_flipper'
 require 'common/exceptions'
 require 'common/client/errors'
 require 'saml/settings_service'
-require 'sentry_logging'
 require 'aes_256_cbc_encryptor'
 
 class ApplicationController < ActionController::API
   include AuthenticationAndSSOConcerns
+  include Error::ErrorHandler
   include SentryLogging
+  include SentryTagging
   include Pundit
-
-  SKIP_SENTRY_EXCEPTION_TYPES = [
-    Common::Exceptions::Unauthorized,
-    Common::Exceptions::RoutingError,
-    Common::Exceptions::Forbidden,
-    Breakers::OutageException
-  ].freeze
 
   VERSION_STATUS = {
     draft: 'Draft Version',
@@ -29,7 +23,6 @@ class ApplicationController < ActionController::API
   prepend_before_action :block_unknown_hosts, :set_app_info_headers
   # Also see AuthenticationAndSSOConcerns for more before filters
   skip_before_action :authenticate, only: %i[cors_preflight routing_error]
-  before_action :set_tags_and_extra_context
 
   def cors_preflight
     head(:ok)
@@ -43,12 +36,6 @@ class ApplicationController < ActionController::API
     InProgressForm.form_for_user(form_id, current_user)&.destroy if current_user
   end
 
-  # I'm commenting this out for now, we can put it back in if we encounter it
-  # def action_missing(m, *_args)
-  #   Rails.logger.error(m)
-  #   raise Common::Exceptions::RoutingError
-  # end
-
   private
 
   attr_reader :current_user
@@ -57,95 +44,6 @@ class ApplicationController < ActionController::API
   def block_unknown_hosts
     return if controller_name == 'example'
     raise Common::Exceptions::NotASafeHostError, request.host unless Settings.virtual_hosts.include?(request.host)
-  end
-
-  def skip_sentry_exception_types
-    SKIP_SENTRY_EXCEPTION_TYPES
-  end
-
-  # rubocop:disable Metrics/BlockLength
-  rescue_from 'Exception' do |exception|
-    # report the original 'cause' of the exception when present
-    if skip_sentry_exception_types.include?(exception.class)
-      Rails.logger.error "#{exception.message}.", backtrace: exception.backtrace
-    else
-      extra = exception.respond_to?(:errors) ? { errors: exception.errors.map(&:to_hash) } : {}
-      if exception.is_a?(Common::Exceptions::BackendServiceException)
-        # Add additional user specific context to the logs
-        if current_user.present?
-          extra[:icn] = current_user.icn
-          extra[:mhv_correlation_id] = current_user.mhv_correlation_id
-        end
-        # Warn about VA900 needing to be added to exception.en.yml
-        if exception.generic_error?
-          log_message_to_sentry(exception.va900_warning, :warn, i18n_exception_hint: exception.va900_hint)
-        end
-      end
-    end
-
-    va_exception =
-      case exception
-      when Pundit::NotAuthorizedError
-        Common::Exceptions::Forbidden.new(detail: 'User does not have access to the requested resource')
-      when ActionController::ParameterMissing
-        Common::Exceptions::ParameterMissing.new(exception.param)
-      when ActionController::UnknownFormat
-        Common::Exceptions::UnknownFormat.new
-      when Common::Exceptions::BaseError
-        exception
-      when Breakers::OutageException
-        Common::Exceptions::ServiceOutage.new(exception.outage)
-      when Common::Client::Errors::ClientError
-        # SSLError, ConnectionFailed, SerializationError, etc
-        Common::Exceptions::ServiceOutage.new(nil, detail: 'Backend Service Outage')
-      else
-        Common::Exceptions::InternalServerError.new(exception)
-      end
-
-    unless skip_sentry_exception_types.include?(exception.class)
-      va_exception_info = { va_exception_errors: va_exception.errors.map(&:to_hash) }
-      log_exception_to_sentry(exception, extra.merge(va_exception_info))
-    end
-
-    headers['WWW-Authenticate'] = 'Token realm="Application"' if va_exception.is_a?(Common::Exceptions::Unauthorized)
-    render json: { errors: va_exception.errors }, status: va_exception.status_code
-  end
-  # rubocop:enable Metrics/BlockLength
-
-  def set_tags_and_extra_context
-    Thread.current['request_id'] = request.uuid
-    Thread.current['additional_request_attributes'] = {
-      'request_ip' => request.remote_ip,
-      'request_agent' => request.user_agent
-    }
-    Raven.extra_context(request_uuid: request.uuid)
-    Raven.user_context(user_context) if current_user
-    Raven.tags_context(tags_context)
-  end
-
-  def user_context
-    {
-      uuid: current_user&.uuid,
-      authn_context: current_user&.authn_context,
-      loa: current_user&.loa,
-      mhv_icn: current_user&.mhv_icn
-    }
-  end
-
-  def tags_context
-    {
-      controller_name: controller_name,
-      sign_in_method: sign_in_method_for_tag
-    }
-  end
-
-  def sign_in_method_for_tag
-    if current_user.present?
-      # account_type is filtered by sentry, becasue in other contexts it refers to a bank account type
-      current_user.identity.sign_in.merge(acct_type: current_user.identity.sign_in[:account_type])
-    else
-      'not-signed-in'
-    end
   end
 
   def set_app_info_headers
