+++ app/controllers/concerns/ignore_not_found.rb
@@ -2,6 +2,6 @@
 
 module IgnoreNotFound
   def skip_sentry_exception_types
-    ApplicationController::SKIP_SENTRY_EXCEPTION_TYPES + [Common::Exceptions::RecordNotFound]
+    super + [Common::Exceptions::RecordNotFound]
   end
 end
